Refactored code into several files
